move copy of vm.args to vm_args config option
I'm going through and submitting pull requests on projects that use relx after a change to the output directory structure. The change is that it is now "releases/<rel_version>/" and not "releases/<rel-name>-<rel_version>", so your copy of vm.args in the overlay would not have worked. But there is also a vm_args option you can use, so I made it use that instead.